requirements (1).txt
streamlit==1.34.0
openai>=1.13.3
Pillow>=10.2.0